requirements.txt
asgiref==3.8.1
Django==5.1.4
django-taggit==6.1.0
python-decouple==3.8
sqlparse==0.5.3
